\name{survPresmooth-package}
\Rdversion{1.1}
\alias{survPresmooth-package}
\alias{survPresmooth}
\docType{package}
\title{Presmoothed Estimation in Survival Analysis}
\description{Compute presmoothed versions of estimators of survival, density, cumulative and non-cumulative hazard functions with right-censored survival data}
\details{
\tabular{ll}{
Package: \tab survPresmooth\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2010-06-01\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Lopez-de-Ullibarri, I. <ilu@udc.es>, Jacome, M. A. <majacome@udc.es>
Maintainer: Lopez-de-Ullibarri, I. <ilu@udc.es>
}


